Visualización sobre más cosas sobre mí con exito y alt de los icons arreglado.
