+++ html/components/global/_base.html
@@ -92,7 +92,9 @@
       gtag('config', 'G-M8H4YB694S');
     </script>
     <script src="../dist/components/particles-js.js"></script>
+    <script type="module" src="../dist/header.js"></script>
     <script type="module" src="../dist/index.js"></script>
+    <script type="module" src="../dist/about-me.js"></script>
     <script type="module" src="../dist/showProyects.js"></script>
   </body>
 </html>
