+++ html/index.html
@@ -97,7 +97,17 @@
         <section class="about-me__presentation">
           <h3 class="about-me__my-name">Hola, soy Galeed Gutierez.</h3>
           <p class="about-me__about-me-text">Soy un joven de 21 años, que vive a los alrededores de una pequeña ciudad de Jujuy, Argentina. Apasionado por la tecnología y la fotografía.</p>
-          <button class="button about-me__button">Leer más</button>
+          <section class="about-me__more-information-container" id="about-me__more-information-container-id">
+            <section class="about-me__fronted-dev">
+              <h3 class="about-me-fronted-dev__title">¿Qué hago como frontend developer?</h3>
+              <p class="about-me-fronted-dev__text">Creo sitios y aplicaciones web, no uso trucos ni mucho menos se trata de magia. Puedo manipular las materias primas de la web, es decir, las principales tecnologías tales como&nbsp;<strong>HTML</strong>,&nbsp;<strong>CSS</strong>,&nbsp;<strong>JavaScript&nbsp;</strong>y&nbsp;<strong>API REST.&nbsp;</strong></p>
+            </section>
+            <section class="about-me__technologies">
+              <h3 class="about-me-technologies__title">Tecnologías que quiero aprender</h3>
+              <p class="what-do-i-do-fronted__second-paragraph">Una de las muchas tecnologías que quiero dominar es, la librería de JavaScript,&nbsp;<strong>React&nbsp;</strong>tanto como para la web como para aplicaciones móviles, un framework de CSS como<strong>Tailwind&nbsp;</strong>o&nbsp;<strong>Boostrap&nbsp;</strong>y sumado a todo esto aprender&nbsp;<strong>NODE.js&nbsp;</strong>para el backend.</p>
+            </section>
+          </section>
+          <button class="button about-me__button" id="about-me__button-id">Leer más</button>
         </section>
       </section>
       <section class="main__portfolio" id="proyects">
@@ -125,7 +135,9 @@
       gtag('config', 'G-M8H4YB694S');
     </script>
     <script src="../dist/components/particles-js.js"></script>
+    <script type="module" src="../dist/header.js"></script>
     <script type="module" src="../dist/index.js"></script>
+    <script type="module" src="../dist/about-me.js"></script>
     <script type="module" src="../dist/showProyects.js"></script>
   </body>
 </html>
