+++ pug/components/global/_scripts.pug
@@ -6,6 +6,8 @@ script.
     gtag('js', new Date());
     gtag('config', 'G-M8H4YB694S');
 script(src="../dist/components/particles-js.js")
+script(type="module" src="../dist/header.js")
 script(type="module" src="../dist/index.js")
+script(type="module" src="../dist/about-me.js")
 script(type="module" src="../dist/showProyects.js")
