+++ sass/components/global/_visibleHidden.scss
@@ -3,3 +3,7 @@
     transform: translateY(27.5rem);
     transition-duration: 1s;
 }
+
+.visibleMoreAboutme {
+    display: block;
+}
