+++ sass/style.scss
@@ -1,7 +1,6 @@
 @import './components/global/global';
 @import './components/global/logo';
 @import './components/header/header';
-@import './components/global/visibleHidden';
 @import './components/icons/icons';
 @import './components/home/home';
 @import './components/global/lineDecorator';
@@ -9,6 +8,7 @@
 @import './components/porfolio/porfolio';
 @import './components/global/icons-social-media';
 @import './components/global/footer';
+@import './components/global/visibleHidden';
