+++ src/showProyects.ts
@@ -9,13 +9,13 @@ const showProyects = () => {
             if (PROYECTS[i].technologies[j] === "PUG") {
                 iconsTechnologies += `
                 <figure class="technologies__icon-figure technologies__pug-icon">
-                    <img id="" src="https://drive.google.com/uc?id=${PROYECTS[i].technologiesIcons[j]}" alt="Icon HTML"/>
+                    <img id="" src="https://drive.google.com/uc?id=${PROYECTS[i].technologiesIcons[j]}" alt="Icon ${PROYECTS[i].technologies[j]}"/>
                 </figure>
             `;
             } else {
                 iconsTechnologies += `
                     <figure class="technologies__icon-figure">
-                        <img id="" src="https://drive.google.com/uc?id=${PROYECTS[i].technologiesIcons[j]}" alt="Icon HTML"/>
+                        <img id="" src="https://drive.google.com/uc?id=${PROYECTS[i].technologiesIcons[j]}" alt="Icon ${PROYECTS[i].technologies[j]}"/>
                     </figure>
                 `;
             }
